check in commit

+++ Sources/structyC/08_two_pointer/is_palindrome.cpp
@@ -9,7 +9,7 @@
 
 bool isPalindrome(std::string s) {
     int l = 0;
-    int r = s.length()-1;
+    int r = (int)s.length() -1 ;
     while (l<r) {
         if (s[l] != s[r]) {
             return false;
